Add BusMapping generation with ordered data sets
This enables the generation of BusMapping instances with
sets of data which are already ordered.
It's not intended to be the final form of the code and can be further
optimized for sure, but pretends to provide a first API so that it's
possible to experiment and play with it.

Resolves: #18