Bump halo2 commit.
This brings in the SimpleFloorPlanner and changes the Circuit trait.

This also updates the public inputs API. (This change will not be
noticeable in this diff since our codebase does not currently use
public inputs.)

This also brings in zero-knowledge blinding factors; however, we
will fork those out in a future commit.