Add `GethQueries` and getBlockByHash fn
Introduce a generic struct which holds a `Provider` (can be Http, Ws,
Ipc) which will be used to make any calls/requests to Geth.

The struct also implements the endpoints that the bus-mapping needs to
query at some point.
In this case, `eth_getBlockByHash` is the one included for now.

Resolves: #23

Co-authored-by: Eduard S. <eduardsanou@posteo.net>