bus-mapping: Introduce EVM error detection
Implement error detection for EVM execution steps.  This PR adds all the
possible errors that can happen during EVM execution, with some of the
error detection incomplete due to missing access to a state trie.  List
of errors as declared by geth:
- Errors not reported
    - [x] ErrInvalidJump
    - [x] ErrReturnDataOutOfBounds
    - [x] ErrExecutionReverted
- Errors ignored
    - [x] ErrCodeStoreOutOfGas
    - [x] ErrMaxCodeSizeExceeded
    - [x] ErrInvalidCode
    - [ ] ErrContractAddressCollision
    - [ ] ErrInsufficientBalance
    - [x] ErrDepth
- Errors reported
    - [x] ErrWriteProtection
    - [x] ErrOutOfGas
    - [x] ErrGasUintOverflow
    - [x] ErrStackUnderflow
    - [x] ErrStackOverflow
    - [x] ErrInvalidOpCode

Each error has a test that uses the external_tracer and produces the
error.

The two incomplete error detections are for
`ErrContractAddressCollision` and `ErrInsufficientBalance`.  Detecting
these errors requires keeping track of the state, which we don't do yet.

Also, bump go-ethereum version in geth-utils.