Refactor test_util for evm_circuits simplifying it
This module was tending to grow into a monster as `mock/lib.rs` was in
the past due to the fact that a lot of functions were being created with
minimal customizations from the generic simple calls.

Instead, the idea would be that on each test the ctx is written. And to
not make it super verbose, `TestContext::simple_ctx_with_bytecode` has
been implemented.

Then, all of the tests that were able to have simpler and easier config
calls have been created.