Add CODEOWNERS file to repo
Since the solution #428 is impossible to implement as a bug affecting
the gh actions prevents it.
An alternative to automatically assign reviews for the reviewing team is
to use a CODEOWNERS file.

So this PR adds this file configured to assign all reviews to
`appliedzkp/zkevm-reviewers` and also Miha specifically to any state
circuit PR.

Resolves: #423