feat: Add specific CODEOWNERS for particular code sections
As discussed in the ZKEVM team meeting at 2022/04/08 we will request a
review extra to mantainers that have build and mantained almost
completely any of the workspace members.

On that way, the person that is most inline with the code is able to
provide a more accurate review for the code changes.