Trigger workflow only when PR is opened
Force-pushing to the PR currently re-triggers the workflow by assigning
to the PR a new member of the team while mantaining the previous.
That's an issue as we end up with multiple reviewers on each PR when
indeed we just need one.

See: https://github.com/uesteibar/reviewer-lottery/issues/26 for more
details.