remove: Delete the review_lottery from the CI workflows
As mentioned in #470, the mechanism was not working due to the
permissions of the user that was creating the PR.

So far, we have explored other alternatives. But for now, since they
would be much more complex to implement it has been decided to not
implement any replacement and simply delete the workflow.

Resolves: #470