Split circuit_input_builder file and improve APIs for it (#482)
* change: Move Error types out from circuit_input_builder

Moved all error type definitions like `OOGError` or `ExecError` to the
`error.rs` file and inter-connected them.

* change: Split CIB file into it's own module

As mentioned in #426, CIB file was already more than 3000 lines of code
and was turning into a deposit of all sorts of code for different
purposes.

This is a preliminary work towards adding a bit of order into the
CIB/bus-mapping crate.

* change: Simplify StepAuxiliaryData struct

Instead of being an enum with repeated fields for Code or CallData
origins, turn it into a single struct which can specify both.

* fix: Add getters to avoid public attrs in structs

* fix: Update codebase to new AuxiliaryData API

* fix: Update codebase to new ReversionGroup API

* fix: Update codebase to new extended Access API

* fix: Adapt Log case for StepAuxData

The latest addition of Log support for ZKEVM in #335 also has needed an
adaptation to the API used to handle the step auxiliary data.

* fix: Apply clippy & rustdoc lints

* Fix BlockContext rwc inc

* fix: Remove getters and use `pub(crate)` when possible

Co-authored-by: Eduard S <eduardsanou@posteo.net>