change: Use single BCC in StateBaseConversion
We were previosly using 25 `BaseConversionConfig`s inside of a single
`StateBaseConversion` config.
Indeed, we can reach the same result and have 25 different regions for
each BaseConversionConfig without the need to have 25 different
instanciations.

This is important since each one of them adds 6 Advice columns and 2
Selector columns.

The after and before sum of columns for the permutation circuits shows a
decent reduction of almost a half of the columns in usage.