change: Move LaneRotateConv fields to `assign_region`
Currently, we were generating 25 `LaneRotateConversionConfig`s inside of
`RhoConfig`.
The only way to prevent this behaviour passes by first refactoring
`LaneRotateConversionConfig` so that it doesn't require `lane` and
`lane_idx` in order to call `configure` and generate `rotation`.

Therefore, the generation of `rotation` has been passed to the witness
assignation step. And also, lane is now not required anymore.

This allows to have 25 regions assigned for the config instead of 25
configs with 1 `Region` each.