change: Reduce RhoConfig gate & column stats
As a consequence of f507d16 and continuing the work on #489 this reduces
the `RhoConfig` to use a simple `LaneRotateConversionConfig` which gets
assigned 25 times witnesses generating 25 different `Region`s instead of
duplicating columns over and over.

This is important since each one of them adds 5 `Advice` columns, 2
`Fixed` columns and 2 Selector columns.