Move keccak inputs computation to CircuitInputBuilder
Previously the TxCircuit and SignVerifyChip had functions to calculate
the keccak inputs that those circuits perform.  I've moved that logic to
the CircuitInputBuilder so that all circuit inputs calculation is done
in a single place.

I've moved some auxiliary types and functions used to generate the
inputs for the SignVerifyChip to eth-types as they are needed to
calculate the SignVerifyChip keccak inputs.

Finally I've added a field in the circuit_input_builder Block to store
the SHA3 opcode inputs so that we don't need to manually reconstruct
them from the Copy Circuit events.

Now the CircuitInputBuilder is able to output the keccak inputs required
by all circuits except for those required by the MPT circuit.

Resolve https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/697