feat: Enable valid MockTx generation
In order to be able to generate Tx objects with full customization
possibilities and, at the same time, have consistent Txs with the
correct hash and signatures, it was required an upgrade on the
`MockTransaction` structure to allow for that.

This includes the following:
- Addition of `AddrOrWallet` enum to keep backwards compatibility and
  allow for simple Tx generations for Opcode testing purposes with just
  addresses or more complex cases where a Wallet is required so that the
  Tx can be correctly signed.
- Refactor of the `MockTransaction` struct making some fields optional
  to preserve the compatibility with all other tests.
- Refactor of the `MockTransaction::build` function to compute the
  signature and tx_hash only when it has not been specified and it is
  possible.

This resolves #677 bringing back the generation of Mock structures for
testing to the mock crate and removing it from any other workspace
member.