change: Use KZG, SHPLONK and Bn256 instead of generics
The first attempt to update the `prover` crate was to basically use full
generics everywhere.

The issue is that we ended up hitting trait safety and the complexity
started to increase significantly.
Hence, since we are always using the same commitment scheme and curve,
it's easier to switch to concrete types.