[WIP] Halo2 24-08-2022 tag update (#724)
* change: Update halo2 tag version in all Cargo.toml

* change: Use `Value` instead of `AssignedCell`.

* remove: Strip keccak circuits from the repo

As decided on
https://github.com/privacy-scaling-explorations/zkevm-circuits/pull/634
we're going to use the bit-approach merged there. And so, most of the
circuits here are no longer required/ will be added back when needed.

* tmp

* fix: RegionCtx new API adoption without &mut holding

* change: Use KZG, SHPLONK and Bn256 instead of generics

The first attempt to update the `prover` crate was to basically use full
generics everywhere.

The issue is that we ended up hitting trait safety and the complexity
started to increase significantly.
Hence, since we are always using the same commitment scheme and curve,
it's easier to switch to concrete types.

* remove: Delete prover crate as was migrated to zkevm-chain

* remove: Remove legacy keccak bench file

* remove: Remove `group` dep from all crates Cargo.toml

* update: Port circuit-benchmarks to latest halo2 version

* update: Port integration-tests to newest halo2 release

* Fix errors in zkevm-circuits tests

* fix: Errors introduced in `main` branch merge

* fix: Load range_chip tables in SignVerigy and Tx Configs

* Fix Synthesis and lookup table errors

* fix: Apply suggestions from @han0110 review

Co-authored-by: Han <tinghan0110@gmail.com>

* Fix Signature errors introduced on Compressed treatment

* fix: Fix tests and address review comments

Co-authored-by: Han <tinghan0110@gmail.com>