Fix: Add dummy advice colum to prevent Phase error
Currently, halo2 prevents the user from adding `Advice<PhaseN>` columns
if no column of the previous phase was defined before.

Since the first thing we need are the `Challenges` and these are placed
with `Phase2` we no longer can compile the code.
Therefore, a dummy column is added so that halo2 does not complain.
This should be solved in the future if the challenge API polishes this.