fix: Mock `Challenges` with different values
As mentioned in #926, the `Challenges::mock` function was not adding
different values for the `Challenges`. Therefore, a new function has
been added that allows mocking by passing two different values (one per
each challenge).

This allows to keep the API we previously had and also, reduce the need
to send always 2 values when only 1 can be passed.

Resolves: #926