fix: Add `extract_field` to utils
I made an effort to migrate `Part<F>` and `PartValue<F>` to use the
`Value<F>` halo2 API. But since the struct doesn't implement `PartialEq`
it's challenging to do it. As I do need to operate with it quite a lot.

Also, there's another caveat which is that it feels really tricky to
extract the byte representation from a `Value<F>`. It feels like if
you're cheating to the API constrantly (using `.map()` to gain access to
the underlying F so that you can do extra things with it such as
decompose it into bytes).

For these reasons, right now `extract_field` was added. An issue will be
filled to track the tech-debt left here and pay it once we maybe we add
the required functionalities to `halo2::Value<F>`.