fix: Introduce padding in the copy_circuit assignation
This includes checks of `CopyDataType::Padding` within the constraints
of the CopyCircuit to enable padding up to
`CircuitsParams::max_copy_rows`.
Also adds assertions to guarantee that the `max_copy_rows` value is
never surpassed during witness assignation.

Finally, it also adds explanations for the `+2` final rows that the
circuit requires due to the `Rotation(2)` usage within it.

Co-authored-by: Eduard S <eduardsanou@posteo.net>