feat: Rework `LookupTable` trait to include annotations
Now the `LookupTable` trait has built-in functionallity to include
annotations for all of the colums of the type that implements the trait.

This has been implemented for all the tables of the zkevm-circuits.
(Internal tables for specific circuits (like byte table in EVM) still
have to be annotated internally (although they implement the trait
automatically).