Remove build.rs in circuit-benchmarks (#1218)
### Description

We remove the build.rs and assign DEGREE in a sane way.

### Issue Link

#1217

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update


### Design choices

I face a design choice to assign the default value for the DEGREE. The
original build.rs assigns them as 19. But in the fix, I can assign
different default values for exp, pi, and tx circuits. I still assign
them 19 to maintain the same behavior.

Note for reviewers. I'm not sure if this breaks the benchmark flow or
introduces dead code to the benchmark script. The `execBench.sh` runs
this line `DEGREE=$k ~/.cargo/bin/cargo test --profile bench
bench_${run_suffix} -p circuit-benchmarks --features benches --
--nocapture > "$target_dir/$logfile" 2>&1`. I believe the `DEGREE=$k`
was used in both the build time for build.rs before and the run time
now. So I expect the bench to work in this fix too.