[Tests] CREATE2 inputs in Math Utils Gadget (#1226)
### Description

The `CREATE2` input length and `RLC(input)` were not tested during the
#1164 merge. This PR adds and enables a test for CREATE2 specific
`input_len` and `input_rlc` calculation.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### How Has This Been Tested?

The helper container can test both `CREATE` and `CREATE2` inputs (length
and `RLC(input_bytes)`). These inputs are then used as inputs to do a
lookup to the Keccak table.

### Related PRs
- #1164 
- #1140 (needs rebasing after this is merged)

---------

Co-authored-by: adria0.eth <5526331+adria0@users.noreply.github.com>