CI: cancel workflow runs of outdated commits and skip for same content (#1163)
Close
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1162

### Summary
Integrate
[skip-duplicate-actions](https://github.com/marketplace/actions/skip-duplicate-actions)
for below CI features:

1. Cancel workflow runs of outdated commits -
[cancel_others](https://github.com/marketplace/actions/skip-duplicate-actions#cancel_others).
e.g. A new commit is added when the CI tasks of old commit has not
finished. With this fix, the old CI tasks should be cancelled
immediately, the new CI tasks should start to run.

2. Skip workflow runs of same content -
[same_content_newer](https://github.com/marketplace/actions/skip-duplicate-actions#skip-concurrent-workflow-runs).
If the same workflow is running on the exact same content (of files),
the new CI runs are skipped.

3. No skip for event `ready_for_review`. Special for zkevm-circuits, CI
of `draft` PRs are not run (configured as [this
code](https://github.com/privacy-scaling-explorations/zkevm-circuits/blob/main/.github/workflows/ci.yml#L15)).
With the above fix `2`, CI tasks are skipped (for same content) and not
exactly run when draft. So CI should always run for ready_for_review
event.

Reference https://github.com/scroll-tech/zkevm-circuits/pull/265 and
https://github.com/scroll-tech/zkevm-circuits/pull/266.