fix account empty (#1237)
### Description

current Account::is_empty is not correct. Fix it to be consistent with
[EIP158](https://eips.ethereum.org/EIPS/eip-158). This will not affect
real world use cases, but i remember this bug causes several testool
cases fail. (These patches are picked from scroll fork, so it is
difficult for me to recall which exact cases will be fixed by this PR
since it is a long time i fixed them...)

### Issue Link

this PR will also fix the (2) problem in
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1118

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- [_item_]

### Rationale

[_design decisions and extended information_]

### How Has This Been Tested?

[_explanation_]

<hr>

## How to fill a PR description 

Please give a concise description of your PR.

The target readers could be future developers, reviewers, and auditors.
By reading your description, they should easily understand the changes
proposed in this pull request.

MUST: Reference the issue to resolve

### Single responsability

Is RECOMMENDED to create single responsibility commits, but not
mandatory.

Anyway, you MUST enumerate the changes in a unitary way, e.g.

```
This PR contains:
- Cleanup of xxxx, yyyy
- Changed xxxx to yyyy in order to bla bla
- Added xxxx function to ...
- Refactored ....
```

### Design choices

RECOMMENDED to:
- What types of design choices did you face?
- What decisions you have made?
- Any valuable information that could help reviewers to think critically