Variadic size tests (#1228)
### Description

Add tests to ensure the circuits size does not depend on the witness.
If this happens it likely means that we are introducing witnesses, or
witness dependent values in fixed columns.

### Issue Link

Close #1108 

### Type of change
- [x] New feature (non-breaking change which adds functionality)

### Contents
Added `variadic_size_check()t` tests, for the following circuits:
- `Copy Circuit`
- `Pi Circuit`
- `State Circuit`
- `Keccak Circuit`
-  `Tx Circuit`
- `EVM Circuit`

### Rationale

The tests have been taken from this PR's (#1064) comments by @CPerezz
and @ed255 and slightly adapted.