Compile time of configure gadget (#1254)
### Description

This improves compilation times. 

### Issue Link

No issue, quality-of-life PR.

### Type of change

- [ x ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

The function `configure_gadget` is split into a generic and a
non-generic functions.

### Rationale

Generics are not free. They are the biggest contributor to compile times
and to CPU code cache. `configure_gadget` is the largest example of this
phenomenon.

I measured about 2x fewer LLVM IR with `cargo llvm-lines`. Even 3x in
Scroll’s fork. The improvement in clock time depends a lot on exactly
what you do, but it will be faster.

### How Has This Been Tested?

    cargo test --features test -p zkevm-circuits --release

Co-authored-by: Aurélien Nicolas <info@nau.re>