subcircuit trait for PiTestCircuit (#1279)
### Description

This implements the `SubCircuit` trait for `PiTestCircuit`.
PiTestCircuit isn't a subcircuit in technical terms and hence panics
when `synthesize_sub` is called with a message to use the `PiCircuit`
for embedding instead.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Rationale

After moving some code of macro bodies I had to update the types and
zkevm-chain now requires the `SubCircuit` trait being implemented for
all circuits used. Technically, PiTestCircuit isn't a Subcircuit. But
without creating a new trait it makes downstream usage more streamlined.

### How Has This Been Tested?

In zkevm-chain

---------

Co-authored-by: Eduard S. <eduardsanou@posteo.net>