add event trigger: workflow_dispatch (#1290)
### Description
With this PR, we enable on demand execution of the integration-tests
workflow, by adding the workflow_dispatch event trigger.

### Type of change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Rationale
Adding the possibility to manually trigger the workflow, acts as an
enabler for integrations tests with real prover. Due to high resource
requirements and extended duration of integration tests with real
prover, it makes sense to not run the wf on each push. Along with
scheduled executions (ie once per week), we provide the option to
manually trigger a parametrized wf execution (select development branch,
select type of prover etc).

* Scheduler and wf parametrization is not part of this PR

```
This PR contains:
- Modification of integration.yml, add the workflow_dispatch event trigger 
```