fix: reduce stack consumption in EVM Circuit (#1294)
### Description

In EVM Circuit's ExecutionConfig, store the gadget for every execution
state in a Box to significantly reduce the size of ExecutionConfig which
inevitably needs to live in the stack at some points. This PR removes
the extended stack setting `RUST_MIN_STACK: 16777216` introduced in
https://github.com/privacy-scaling-explorations/zkevm-circuits/pull/1286

Let's see if this works!

### Issue Link

Related to
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1287

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Rationale

Even though setting the env var `RUST_MIN_STACK=16777216` fixes the
stack overflow we encountered before, I think it's better to resolve
this by avoiding consuming so much stack that it overflows on its
default size. This is because setting this env var can be annoying and
it's easy to forget it in many scenarios.