add scheduler, add conditional parameter provertype, add workflow_dis… (#1297)
### Description
As a follow up to #1290, we add required input choices for manual
trigger (workflow_dispatch), as a means to define whether we want to
invoke mock or real prover and select target branch.
We introduce environment variable [provertype] 
We also add a scheduler to trigger integration tests every Sunday at
01:50AM. Scheduled job will default to real_prover.
Consecutiveness job is also now enforced on scheduled events and
workflow_dispatch

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- Changed File: .github/workflows/integration.yml

### Rationale
Overall, we want integration tests to be invoked on 4 distinct
occasions:
1. Ready for Review PRs, labelled as 'trigger-integration-tests'
2. On branch push to main
3. On a predefined interval (once a week on offline hours for now)
4. On demand

1 and 2 are always executed with mock object to save time and cloud
resources
3 is always executed with real_prover
4 allows selection of both prover type and branch to execute integration
tests against.

### Design choices

Due to gh actions not supporting ternary operations on environment
variables
for example:
    if github.event == pull_request:
        env.provertype="mock_prover"
    elif github.event == schedule:
        env.provertype="real_prover"

we introduce new job steps to conditionally inject ${{ env.provertype }}
to $GITHUB_ENV, for example:
    - name: Set provertype to workflow_dispatch choice
        run: |
          echo "provertype=${{ inputs.provertype }}" >> $GITHUB_ENV
        if: github.event_name == 'workflow_dispatch'