Integration tests with ephemeral runners via GitHub ops (#1319)
### Description

This PR introduces github-ops functionality to
zkevm-circuits/integration-tests, allowing execution on ephemeral AWS
hosted runners. Execution mode is parameterized based on event trigger
type and prover type selection.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- .github/workflows/integration.yml
- integration-tests/run.sh
- integration-tests/README.md

```
This PR contains:
- A new job that initializes prover and instance type workflow variables 
- run.sh updated to Compose V2
- Updated README.md file to reflect the updated docker compose commands
```

### Design choices
As per new integration tests execution mode, we run the tests manually,
on schedule or on specific GitHub events (PR, push to main). Depending
on workflow setup, test is executed either with real or mock prover,
thus different AWS instance is selected to accommodate increased
resource requirements for real prover. For this reason, we introduced a
new job that initializes instance type and prover type variables as job
outputs to be used for subsequent jobs (integration-tests).

---------

Co-authored-by: testuser-at-490752553772 <m_aronis@hotmail.com>