Fix `ExecutionState::halts_in_exception` (#1299)
### Description

This PR aims to remove `ExecutionState::{ErrorDepth,InsufficientBalance
,ErrorContractAddressCollision}` from
`ExecutionState::halts_in_exception` and handle `InsufficientBalance`
properly by `CallOpGadget`

### Issue Link


https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1298

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

Some exception states happen right in the moment when we are doing
`*CALL*` or `CREATE*`, including:

- `ErrorDepth`
- `ErrorInsufficientBalance`
- `ErrorContractAddressCollision`

Which will only be treated as failure of sub-call, and won't halt
current call.

```mermaid
stateDiagram-v2
    state fork <<fork>>
    state join <<join>>
    state NextCall {
        ...* --> join
    }
    state CurrentCall {
        ... --> fork
        fork --> CALL|CREATE
        fork --> ErrorDepth
        fork --> ErrorInsufficientBalance
        fork --> ErrorContractAddressCollision
        CALL|CREATE --> NextCall
        ErrorDepth --> join
        ErrorInsufficientBalance --> join
        ErrorContractAddressCollision --> join
        join --> NextStep
    }
```

However, if we want to handle these as a separate state we'd have many
constraint duplicated (e.g. gas cost calculation), so
`ErrorInsufficientBalance` is merged into `CALL` in
https://github.com/privacy-scaling-explorations/zkevm-circuits/pull/998
without too much overhead. For the other 2 states we could apply the
same strategy to avoid confusion.


### Rationale

### How Has This Been Tested?

### Design choices