[refactor] refactor and add more helper in Bytecode (#1365)
### Description

Refactor and add more helper functions to Bytecode:
- Removed several functions such as `call`, `balance`, `mstore`,
`calldatacopy`, and `return_bytecode`
- Added `op_jumpdest` function
- Implemented `impl_push_n` macro to generate functions for `op_push1`
to `op_push32`
- Implemented `impl_other_opcodes` macro to generate functions for
various opcodes like `op_stop`, `op_add`, `op_mul`, `op_sub`, and so on
- Renamed opcode helper functions to `op_{opcode}` for consistency and
clarity

### Issue Link

N/A

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- Removal of several functions
- Addition of `op_jumpdest` function
- Implementation of `impl_push_n` macro to generate functions for
`op_push1` to `op_push32`
- Implementation of `impl_other_opcodes` macro to generate functions for
various opcodes
- Renaming of opcode helper functions

### Rationale

This refactor improves the code by removing unused functions, adding a
new function, and implementing macros to generate functions for opcodes.
The renaming of the opcode helper functions to `op_{opcode}` provides
consistency and clarity to the codebase.

### How Has This Been Tested?

https://github.com/scroll-tech/zkevm-circuits/actions/runs/4751119765