Add `SubCircuit::unusable_rows` and tests for each `SubCircuit` (#1282)
### Description

To remove magic numbers for blinding factors, this PR adds an associated
function `unusable_rows` for `SubCircuit` to returns their
`unusable_rows`, which should be `meta.blinding_factors() + 1`.

Tests are also added to make sure the returned values are correct. 

Note that currently `KeccakCircuit` could be configured by environment
variable `KECCAK_ROWS` to decide how many rows a round would take, which
affect the different rotation queried, and it's unfortunately not a easy
way to compute how many different rotations are used, so this PR
hardcodes the unusable rows from `KECCAK_ROWS = 1` to `20` for easy
lookup.

### Issue Link


https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/949

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- [_item_]

### Rationale

[_design decisions and extended information_]

### How Has This Been Tested?

[_explanation_]

---------

Co-authored-by: David Nevado <davidnevadoc@users.noreply.github.com>
Co-authored-by: "adria0.eth" <5526331+adria0@users.noreply.github.com>
Co-authored-by: Eduard S <eduardsanou@posteo.net>