Patch halo2 dependency
Path to a recent commit from halo2 that contains a removal of an aggresive
check that caused many tests to fail, eventhough the tests and the circuits are
OK.  See https://github.com/privacy-scaling-explorations/halo2/pull/179

Cargo doesn't support patching a git repository to point to the same repository
with different branch/rev/tag.  So I'm abusing the fact that previously our
github organization was `appliedzkp` to trick Cargo into thinking it's a
different git repo (but it's just an alias from github).