Update halo2 dependency to v2023_04_20 (#1374)
### Description

Update halo2 to the latest dependency, which contains some changes
affecting the `ff` trait.

### Issue Link

Resolve #1366

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update


### Rationale

I've extended the `Field` trait we define in `eth_types` to contain the
common traits used in different places, to avoid having a custom list of
traits in each function.