Fix infinite loop (#1421)
### Description

Fix an infinite loop from a `From` implementation calling itself.

### Issue Link

Resolve
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1420
Related
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1417

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

### Contents

@ChihChengLiang found that this issue was introduced in
https://github.com/privacy-scaling-explorations/zkevm-circuits/pull/1406
and by debugging on the debug build I found the infinite loop (which
caused the stack to overflow which terminated the program with a
segfault). The segfault is not observed in the release build, I guess
the compiler optimizes the recursive call into an infinite loop, so the
stack is never exhausted.