Feat/CREATE Part B - Error cases that will be handled within opcode gadgets (#1425)
### Description

**NOTE: This is an updated version of
https://github.com/privacy-scaling-explorations/zkevm-circuits/pull/1357**

This PR is actually based on top of #1419

### Issue Link

#1130 

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents
Error types for insufficient balance and nonce overflow. These are
supporting changes required for the CREATE/CREATE2 opcodes' gadget.

### Rationale
The above errors will be handled within the CREATE/CREATE2 opcodes'
gadget. In case of insufficient balance, it is also handled in the
CallOp gadget for call related opcodes.

---------

Co-authored-by: Rohit Narurkar <rohit.narurkar@protonmail.com>